requirements.txt
opencv-python
psycopg2
matplotlib
kafka-python
docker
